fix sprite collision

+++ srcs/display.c
@@ -6,7 +6,7 @@
 /*   By: eassouli <eassouli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/16 16:57:15 by eassouli          #+#    #+#             */
-/*   Updated: 2020/10/28 16:39:47 by eassouli         ###   ########.fr       */
+/*   Updated: 2020/10/28 21:01:45 by eassouli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void	draw_walls(t_a *a)
 	{
 		pos = (a->img.px_start * a->img.size_l + a->mlx.x * (a->img.bpp / 8));
 		a->tex.y = (int)a->tex.pos & (a->tex.dim - 1);
-		*(unsigned int *)(a->img.img + pos) = tx.img[a->tex.y *
+		*(unsigned int *)(a->img.img + pos) = tx.img[(int)a->tex.y *
 		(tx.size_l / 4) + a->tex.x];
 		a->tex.pos += a->tex.step;
 		a->img.px_start++;
@@ -78,7 +78,7 @@ void	draw_sprite(t_a *a)
 				pos = (y * a->img.size_l + stripe * (a->img.bpp / 8));
 				d = y * 256 - a->res.h * 128 + a->spr.spriteHeight * 128;
 				a->tex.y = ((d * a->tex.dim) / a->spr.spriteHeight) / 256;
-				color = a->spr_txr.img[a->tex.y * (a->spr_txr.size_l / 4) + a->tex.x];
+				color = a->spr_txr.img[(int)a->tex.y * (a->spr_txr.size_l / 4) + a->tex.x];
 				if ((color & 0xFFFFFF) != 0)
 					*(unsigned int *)(a->img.img + pos) = color;
 				y++;
